Renamed unclear /dash/ page to /views/
