+++ src/routes.js
@@ -3,13 +3,13 @@ import LoginView from './components/Login.vue'
 import NotFoundView from './components/404.vue'
 
 // Import Views - Dash
-import DashboardView from './components/dash/Dashboard.vue'
-import TablesView from './components/dash/Tables.vue'
-import TasksView from './components/dash/Tasks.vue'
-import SettingView from './components/dash/Setting.vue'
-import AccessView from './components/dash/Access.vue'
-import ServerView from './components/dash/Server.vue'
-import ReposView from './components/dash/Repos.vue'
+import DashboardView from './components/views/Dashboard.vue'
+import TablesView from './components/views/Tables.vue'
+import TasksView from './components/views/Tasks.vue'
+import SettingView from './components/views/Setting.vue'
+import AccessView from './components/views/Access.vue'
+import ServerView from './components/views/Server.vue'
+import ReposView from './components/views/Repos.vue'
 
 // Routes
 const routes = [
